pyproject.toml
[tool.poetry]
name = "discord-bluemchen"
version = "0.1.0"
description = "bluemchen 🌻"
authors = ["Moritz Klein <moritz.klein@pm.me>"]
license = "MIT"
readme = "README.md"
# packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
